Projeto Imigrante - Backend implementado em sua completude

+++ src/main/java/br/com/projetoimigrante/api/model/Regiao.java
@@ -1,5 +1,6 @@
 package br.com.projetoimigrante.api.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -16,7 +17,7 @@ public class Regiao {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 
 	private String nome;
